see_run_desk
Former-commit-id: 2a68cb631d5135702e5599d05ce11b8f520f4119

+++ utils/util.py
@@ -33,11 +33,12 @@ def get_file_list(dir):
     return file_names
 
 
-def clean_txt():
+def clean_txt(clean=True):
     tmp_files = os.listdir("./tmp")
-    for name in tmp_files:
-        if ".txt" in name:
-            os.remove("./tmp/"+name)
+    if clean:
+        for name in tmp_files:
+            if ".txt" in name:
+                os.remove("./tmp/"+name)
     return tmp_files
